refactor(cli): split pr.rs + pr_sync.rs under 300-line cap
CI file-size guard caught the violation (pr.rs 316, pr_sync.rs
349 vs 300 cap). Pure refactor: extract pure parsers into
sibling modules.

## Problem

Adding `cvg pr sync` (T2.04) pushed `pr.rs` to 316 lines and
`pr_sync.rs` to 349 — both over the 300-line Rust cap that CI
enforces in the file-size guard step.

## Why

CONSTITUTION § 13 cap. Splitting also matches the pattern
already used for `pr_diff.rs` and `pr_render.rs`: orchestration
in the namesake file, pure helpers in siblings.

## What changed

- New `pr_parse.rs`: `ParsedManifest` + `parse_manifest()` +
  3 unit tests, extracted from `pr.rs`.
- New `pr_sync_parse.rs`: `parse_tracks_lines()` +
  `is_valid_uuid()` + 10 unit tests, extracted from
  `pr_sync.rs`.
- `pr.rs` now imports `parse_manifest`/`ParsedManifest` from
  `pr_parse`. `pr_diff.rs` updated likewise.
- `pr_sync.rs` now imports `parse_tracks_lines` from
  `pr_sync_parse`.
- `commands/mod.rs` declares both new sibling modules.

No public API change. No behaviour change.

## Validation

- `cargo fmt --all -- --check` clean.
- `cargo clippy --workspace --all-targets -D warnings` clean.
- `cargo test -p convergio-cli pr_`: **21 + 3 passed, 0 failed**
  (unit tests on pr_sync_parse + pr_parse + pr_render +
  cli integration tests).
- `wc -l crates/convergio-cli/src/commands/pr*.rs`: every file
  ≤ 244 lines, well under the 300 cap.

## Impact

- Closes the CI file-size-guard failure on PR #59.
- Matches the existing split pattern (`pr_diff.rs`,
  `pr_render.rs`) — no surprise for future readers.

## Files touched

- `crates/convergio-cli/src/commands/pr.rs` (316 → 231)
- `crates/convergio-cli/src/commands/pr_sync.rs` (349 → 244)
- `crates/convergio-cli/src/commands/pr_diff.rs` (import path)
- `crates/convergio-cli/src/commands/pr_parse.rs` (new, 96)
- `crates/convergio-cli/src/commands/pr_sync_parse.rs` (new, 106)
- `crates/convergio-cli/src/commands/mod.rs` (+ 2 mod decls)
- `docs/INDEX.md` (auto-regenerated)

Tracks: 5253cdd6-425d-4210-9173-efc3d8c4ec5e
