+++ crates/convergio-cli/src/commands/mod.rs
@@ -23,8 +23,10 @@ pub mod mcp;
 pub mod plan;
 pub mod pr;
 mod pr_diff;
+mod pr_parse;
 mod pr_render;
 mod pr_sync;
+mod pr_sync_parse;
 pub mod service;
 pub mod session;
 mod session_render;
